relocated the cryptor
